+++ src/main/java/edu/cnm/deepdive/join_me_backend/model/dao/SquareRepository.java
@@ -4,7 +4,7 @@ import edu.cnm.deepdive.join_me_backend.model.entity.Square;
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SquareRepository extends CrudRepository<Square, Long> {
+public interface SquareRepository extends CrudRepository<Square, Integer> {
 
   List<Square> findAllByOrderByIdAsc();
