+++ src/main/java/edu/cnm/deepdive/join_me_backend/model/dao/VertexRepository.java
@@ -4,7 +4,7 @@ import edu.cnm.deepdive.join_me_backend.model.entity.Vertex;
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 
-public interface VertexRepository extends CrudRepository<Vertex, Long> {
+public interface VertexRepository extends CrudRepository<Vertex, Integer> {
 
   List<Vertex> findAllByOrderByIdAsc();
